# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/phat": {
      "get": {
        "description": "Phiếu phạt do DB tự lập khi trả sách quá hạn/hư hỏng/mất. Bạn đọc xem phạt của mình ở `GET /me/tien-phat`.\n\n**Quyền:** `ADMIN`, `THU_THU`",
        "operationId": "Phat_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Trang, bắt đầu từ 1.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số dòng mỗi trang (1–100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "trangThai",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CHUA_THANH_TOAN",
                "DA_THANH_TOAN",
                "HUY"
              ]
            }
          },
          {
            "name": "maNguoiDung",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PhieuPhatChiTietDto"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Danh sách phiếu phạt (phân trang, lọc theo trạng thái/người dùng)",
        "tags": [
          "phat"
        ],
        "x-roles": [
          "ADMIN",
          "THU_THU"
        ]
      }
    },
    "/phat/{id}/thanh-toan": {
      "post": {
        "description": "Qua `sp_thanh_toan_phat`. Phiếu không tồn tại hoặc không ở trạng thái `CHUA_THANH_TOAN` thì 422.\n\n**Quyền:** `ADMIN`, `THU_THU`",
        "operationId": "Phat_thanhToan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuPhatDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thanh toán phiếu phạt",
        "tags": [
          "phat"
        ],
        "x-roles": [
          "ADMIN",
          "THU_THU"
        ]
      }
    },
    "/phat/{id}/huy": {
      "post": {
        "description": "Qua `sp_huy_phat` (bắt buộc ghi lý do). Chỉ ADMIN; phiếu không ở trạng thái `CHUA_THANH_TOAN` thì 422.\n\n**Quyền:** `ADMIN`",
        "operationId": "Phat_huy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HuyPhatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhieuPhatDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu gửi lên không hợp lệ (message là mảng lỗi validate) hoặc vi phạm CHECK của CSDL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Vai trò của tài khoản không được phép gọi endpoint này",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bản ghi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Vi phạm quy tắc nghiệp vụ do procedure/trigger của CSDL báo (message là thông báo của CSDL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hủy phiếu phạt",
        "tags": [
          "phat"
        ],
        "x-roles": [
          "ADMIN"
        ]
      }
    },
...
    "/me/tien-phat": {
      "get": {
        "description": "Mọi phiếu phạt (kể cả đã thanh toán/hủy), mới nhất trước. Thanh toán do thủ thư thực hiện ở `POST /phat/{id}/thanh-toan`.\n\n**Quyền:** mọi tài khoản đã đăng nhập.",
        "operationId": "BanDoc_tienPhat",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TienPhatCuaToiDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Thiếu/sai/hết hạn JWT, hoặc tài khoản/người dùng đang bị khóa hay ngừng hoạt động",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tiền phạt của tôi",
        "tags": [
          "ban-doc"
        ]
      }
    },
...
  },
  "info": {
    "title": "QLTV Nhóm 8 - API quản lý thư viện",
    "description": "API quản lý thư viện đại học (đồ án IE103, nhóm 8). Nghiệp vụ nằm trong MySQL (procedure, trigger, view); BE chỉ gọi lại và xác thực.\n\n**Cách dùng nhanh**\n1. `POST /auth/login` với tài khoản seed (ví dụ `ad001` / `AD001@Nhom8`), lấy `accessToken`.\n2. Bấm **Authorize** ở góc phải và dán token (không cần gõ chữ `Bearer`). Token được nhớ khi tải lại trang.\n3. Mỗi endpoint ghi rõ **Quyền** được gọi. Sai vai trò: 403. Token sai, hết hạn hoặc tài khoản bị khóa: 401.\n\n**Quy ước dữ liệu**\n- Khóa chính BIGINT (`id`) là **chuỗi**, ví dụ `\"13\"`. Các mã (`maSach`, `maBanSach`, `maNguoiDung`...) là chuỗi và là thứ dùng khi gọi API nghiệp vụ.\n- Endpoint danh sách có phân trang trả `{ data, total, page, limit }` (`page` từ 1, `limit` tối đa 100).\n- Cột ngày kiểu DATE trả dạng ISO `YYYY-MM-DDT00:00:00.000Z`.\n- Tiền (VND) từ báo cáo và `/me` là number; `soTien`/`giaBia` đọc từ bảng là chuỗi thập phân (Prisma Decimal).\n- Báo cáo (`/bao-cao/*`), `GET /sach` và `/me/*` giữ tên cột snake_case của view; các endpoint còn lại dùng camelCase.\n- Lỗi luôn có dạng `{ statusCode, message }`. 422 là lỗi nghiệp vụ do CSDL báo, `message` là thông báo cho người dùng (tiếng Việt không dấu).",
    "version": "2.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "error": {
            "type": "string",
            "example": "Unprocessable Entity",
            "description": "Tên mã HTTP (chỉ có ở lỗi do bộ lọc CSDL và một số lỗi của Nest)"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Nguoi dung dang no phat",
            "description": "Chuỗi thông báo; lỗi validate dữ liệu gửi lên (400) là mảng các thông báo"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "NguoiDungTomTatDto": {
        "type": "object",
        "properties": {
          "maNguoiDung": {
            "type": "string",
            "example": "SV001"
          },
          "hoTen": {
            "type": "string",
            "example": "Võ Hoàng Nhiên"
          }
        },
        "required": [
          "maNguoiDung",
          "hoTen"
        ]
      },
...
      "LoaiPhat": {
        "type": "string",
        "enum": [
          "QUA_HAN",
          "HU_HONG",
          "MAT_SACH"
        ]
      },
      "TrangThaiPhieuPhat": {
        "type": "string",
        "enum": [
          "CHUA_THANH_TOAN",
          "DA_THANH_TOAN",
          "HUY"
        ]
      },
      "PhieuPhatDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "Mã phiếu phạt (dùng ở /phat/:id/...)",
            "nullable": false
          },
          "ctPhieuMuonId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "loaiPhat": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/LoaiPhat"
              }
            ]
          },
          "soTien": {
            "type": "string",
            "example": "85000.00",
            "description": "Số tiền phạt (VND)",
            "nullable": false
          },
          "lyDo": {
            "type": "string",
            "nullable": true
          },
          "trangThai": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/TrangThaiPhieuPhat"
              }
            ]
          },
          "ngayTao": {
            "type": "string",
            "format": "date-time"
          },
          "ngayThanhToan": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ctPhieuMuonId",
          "loaiPhat",
          "soTien",
          "lyDo",
          "trangThai",
          "ngayTao",
          "ngayThanhToan"
        ]
      },
...
      "PhatSachDto": {
        "type": "object",
        "properties": {
          "tenSach": {
            "type": "string"
          }
        },
        "required": [
          "tenSach"
        ]
      },
      "PhatBanSachDto": {
        "type": "object",
        "properties": {
          "maBanSach": {
            "type": "string",
            "example": "BS001"
          },
          "sach": {
            "$ref": "#/components/schemas/PhatSachDto"
          }
        },
        "required": [
          "maBanSach",
          "sach"
        ]
      },
      "PhatPhieuMuonDto": {
        "type": "object",
        "properties": {
          "maPhieu": {
            "type": "string",
            "example": "PM001"
          },
          "nguoiDung": {
            "$ref": "#/components/schemas/NguoiDungTomTatDto"
          }
        },
        "required": [
          "maPhieu",
          "nguoiDung"
        ]
      },
      "PhatCtPhieuMuonDto": {
        "type": "object",
        "properties": {
          "banSach": {
            "$ref": "#/components/schemas/PhatBanSachDto"
          },
          "phieuMuon": {
            "$ref": "#/components/schemas/PhatPhieuMuonDto"
          }
        },
        "required": [
          "banSach",
          "phieuMuon"
        ]
      },
      "PhieuPhatChiTietDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "Mã phiếu phạt (dùng ở /phat/:id/...)",
            "nullable": false
          },
          "ctPhieuMuonId": {
            "type": "string",
            "example": "1",
            "nullable": false
          },
          "loaiPhat": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/LoaiPhat"
              }
            ]
          },
          "soTien": {
            "type": "string",
            "example": "85000.00",
            "description": "Số tiền phạt (VND)",
            "nullable": false
          },
          "lyDo": {
            "type": "string",
            "nullable": true
          },
          "trangThai": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/TrangThaiPhieuPhat"
              }
            ]
          },
          "ngayTao": {
            "type": "string",
            "format": "date-time"
          },
          "ngayThanhToan": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "ctPhieuMuon": {
            "$ref": "#/components/schemas/PhatCtPhieuMuonDto"
          }
        },
        "required": [
          "id",
          "ctPhieuMuonId",
          "loaiPhat",
          "soTien",
          "lyDo",
          "trangThai",
          "ngayTao",
          "ngayThanhToan",
          "ctPhieuMuon"
        ]
      },
      "HuyPhatDto": {
        "type": "object",
        "properties": {
          "lyDo": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "lyDo"
        ]
      },
...
      "TienPhatCuaToiDto": {
        "type": "object",
        "properties": {
          "ma_phieu_phat": {
            "type": "string",
            "example": "1",
            "description": "Mã phiếu phạt",
            "nullable": false
          },
          "loai_phat": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/LoaiPhat"
              }
            ]
          },
          "so_tien": {
            "type": "number",
            "description": "VND"
          },
          "ly_do": {
            "type": "string",
            "nullable": true
          },
          "trang_thai": {
            "nullable": false,
            "allOf": [
              {
                "$ref": "#/components/schemas/TrangThaiPhieuPhat"
              }
            ]
          },
          "ngay_tao": {
            "type": "string",
            "format": "date-time",
            "description": "Kiểu DATE"
          },
          "ngay_thanh_toan": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "ma_phieu_phat",
          "loai_phat",
          "so_tien",
          "ly_do",
          "trang_thai",
          "ngay_tao",
          "ngay_thanh_toan"
        ]
      },
...
    }
  }
}